moving transaction to controllers as it best belongs there
